+++ lib/src/v2/view/scene_view_interactive_v2.dart
@@ -133,6 +133,7 @@ class _SceneViewInteractiveV2State extends State<SceneViewInteractiveV2> {
           textLayoutCache: _textLayoutCache,
           strokePathCache: _strokePathCache,
           pathMetricsCache: _pathMetricsCache,
+          selectionRect: widget.controller.selectionRect,
           selectionColor: widget.selectionColor,
           selectionStrokeWidth: widget.selectionStrokeWidth,
           gridStrokeWidth: widget.gridStrokeWidth,